Implements slices 01, 02, 03, 04, 05 (backend) and 11 of the launch handoff, plus
the shared foundation and phase-2 infrastructure, and brings the CI merge gate to
green.

Highlights:
- packages/contracts: canonical event envelope, generated state machines, money,
  city config, flags, error codes, testIDs.
- A runnable Prisma migration chain (was unvalidatable and un-appliable) with a
  double-entry ledger invariant enforced by the database.
- Slice 02 Move lockstep (Go ride-service), 67 tests, race-clean.
- Slice 03 identity + signed gateway context, 281 tests.
- Slice 04 wallet double-entry ledger + finance recon, 85 tests.
- Slice 11 support/safety/audit, 55 tests.
- Slice 05 Bites backend, outbox relay, config/flags, Lagos config.
- payment-service typechecks against real Prisma types (any-shim removed);
  superseded legacy quarantined and documented.
- CI: typecheck 36/36, merge gate (build/test/go-build/db-check/contracts) green.

Residual, documented in docs/launch-readiness: eslint legacy debt (~1,400,
pre-existing), mobile apps not analyze-clean, and PSP collection / payout batches
/ payment webhooks not yet on the canonical ledger.

+++ mobile/packages/core/lib/src/domain/usecases/ride_usecases.dart
@@ -3,7 +3,7 @@
 /// Business logic for ride operations.
 library;
 
-import '../../../core.dart';
+import '../../../ubi_core.dart';
 
 /// Get ride fare estimates
 class GetRideEstimatesUseCase implements UseCase<List<RideEstimate>, GetRideEstimatesParams> {
@@ -124,7 +124,7 @@ class RateRideParams {
   });
 
   final String rideId;
-  final int rating;
+  final double rating;
   final String? review;
 }
